+++ src/gql/interfaces/EmbeddedAsset.php
@@ -216,6 +216,18 @@ class EmbeddedAsset extends Element
                 'type' => Type::string(),
                 'description' => 'The embed HTML for the embedded asset.'
             ],
+            'iframeSrc' => [
+                'name' => 'iframeSrc',
+                'type' => Type::string(),
+                'description' => 'The embed URL for the embedded asset, if the embedded asset\'s code is an iframe.',
+                'args' => [
+                    'params' => [
+                        'name' => 'params',
+                        'type' => Type::nonNull(Type::listOf(Type::string())),
+                        'description' => 'The parameters to add to the embed URL.',
+                    ],
+                ],
+            ],
         ], self::getName());
     }
 }