+++ src/gql/types/EmbeddedAsset.php
@@ -3,6 +3,7 @@
 namespace spicyweb\embeddedassets\gql\types;
 
 use craft\gql\types\elements\Element;
+use GraphQL\Type\Definition\ResolveInfo;
 use spicyweb\embeddedassets\gql\interfaces\EmbeddedAsset as EmbeddedAssetInterface;
 
 /**
@@ -21,5 +22,16 @@ class EmbeddedAsset extends Element
         $config['interfaces'] = [EmbeddedAssetInterface::getType()];
 
         parent::__construct($config);
+    }
+
+    /**
+     * @inheritdoc
+     * @since 2.7.0
+     */
+    protected function resolve($source, $arguments, $context, ResolveInfo $resolveInfo)
+    {
+        return $resolveInfo->fieldName === 'iframeSrc'
+            ? $source->{'get' . ucfirst($resolveInfo->fieldName)}($arguments['params'])
+            : parent::resolve($source, $arguments, $context, $resolveInfo);
     }
 }